+clients +music support +fixes

+++ src/GachaMoon.Domain/ExternalServices/UserAnimeData.cs
@@ -1,6 +1,6 @@
 namespace GachaMoon.Domain.ExternalServices;
 public class UserAnimeData
 {
-    public int Id { get; set; } = default!;
+    public long Id { get; set; } = default!;
     public long InternalId { get; set; } = default!;
 }
